requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
sqlalchemy==1.4.46
pydantic==1.10.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0